Rename the package namespace
